feat: 도서 평점 CUD[#S12P21B203-79]

See merge request s12-bigdata-recom-sub1/S12P21B203!45

+++ backend/booknest/src/main/java/com/ssafy/booknest/domain/book/repository/ebookRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ebookRepository extends JpaRepository<Ebook, Long> {
+public interface ebookRepository extends JpaRepository<Ebook, Integer> {
 
     List<Ebook> findByCityAndDistrict(String city, String district);
 }
